+++ src/inspect_ai/_util/content.py
@@ -1,8 +1,11 @@
-from typing import Literal, Sequence, Union
+import mimetypes
+from pathlib import Path
+from typing import Any, Literal, Sequence, Union
 
-from pydantic import BaseModel, Field, JsonValue
+from pydantic import BaseModel, Field, JsonValue, model_validator
 
 from inspect_ai._util.citation import Citation
+from inspect_ai._util.url import data_uri_mime_type
 
 
 class ContentBase(BaseModel):
@@ -109,10 +112,59 @@ class ContentVideo(ContentBase):
     """Type."""
 
     video: str
-    """Audio file path or base64 encoded data URL."""
+    """Video file path or base64 encoded data URL."""
 
     format: Literal["mp4", "mpeg", "mov"]
     """Format of video data ('mp4', 'mpeg', or 'mov')"""
+
+
+class ContentDocument(ContentBase):
+    """Document content (e.g. a PDF)."""
+
+    type: Literal["document"] = Field(default="document")
+    """Type."""
+
+    document: str
+    """Document file path or base64 encoded data URL."""
+
+    filename: str = Field(default_factory=str)
+    """Document filename (automatically determined from 'document' if not specified)."""
+
+    mime_type: str = Field(default_factory=str)
+    """Document mime type (automatically determined from 'document' if not specified)."""
+
+    @model_validator(mode="before")
+    @classmethod
+    def set_name_and_mime_type(cls, data: dict[str, Any]) -> dict[str, Any]:
+        """Automatically set name and mime_type if not provided."""
+        document: str | None = data.get("document")
+        filename: str | None = data.get("filename")
+        mime_type: str | None = data.get("mime_type")
+
+        if not document:
+            # Let Pydantic handle the missing required field
+            return data
+
+        if document.startswith("data:"):
+            if not mime_type:
+                mime_type = data_uri_mime_type(document) or "application/octet-stream"
+            if not filename:
+                extension = mime_type.split("/")[-1]
+                filename = f"document.{extension}"
+
+        else:
+            path = Path(document)
+            if not filename:
+                filename = path.name
+
+            if not mime_type:
+                guessed_type, _ = mimetypes.guess_type(str(path))
+                mime_type = guessed_type or "application/octet-stream"
+
+        data["filename"] = filename
+        data["mime_type"] = mime_type
+
+        return data
 
 
 class ContentData(ContentBase):
@@ -133,5 +185,6 @@ Content = Union[
     ContentVideo,
     ContentData,
     ContentToolUse,
+    ContentDocument,
 ]
 """Content sent to or received from a model."""