+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
-- MCP: Support for [Remote MCP Servers](https://inspect.aisi.org.uk/tools-mcp.html#remote-mcp) for OpenAI and Anthropic models.
+- Support for [PDF input](https://inspect.aisi.org.uk/multimodal.html#pdf) for OpenAI, Anthropic, and Google via new `ContentDocument` type.
+- MCP: Use [Remote MCP Servers](https://inspect.aisi.org.uk/tools-mcp.html#remote-mcp) with OpenAI and Anthropic models.
 - OpenAI: Use types from latest SDK (v1.99.7) and make that the minimum required version of the `openai` package.
 - OpenAI: Automatically use background-mode for deep research models.
 - Scoring: NaN values returned from scorers will be excluded from reductions when reducing epochs.